Fix apprentice dashboard crash: move nav-config helper to server-safe module
The /apprentice layout (a server component) imported and called
resolveApprenticeNavConfig from ApprenticeSubNav.tsx, which is a
'use client' module. Next.js forbids invoking a client-module function
on the server, so every /apprentice* page crashed with 'Attempted to
call resolveApprenticeNavConfig() from the server'. Move the pure helper
to lib/portal/apprentice-nav-config.ts (server-safe; reads the
server-safe APPRENTICE_PORTAL_CONFIGS) and import it there.
